Move row limit fns to own helper
