+++ server/api/[table]/alerts.ts
@@ -10,7 +10,8 @@ import {
 } from "@/server/dataProcessing/dataFilters";
 import { buildMinimalFeatureCollection } from "@/utils/geoUtils";
 import { validatePermissions } from "@/utils/accessControls";
-import { parseBasemaps, parseAndValidateLimit } from "@/server/utils";
+import { parseBasemaps } from "@/server/utils";
+import { parseAndValidateLimit } from "@/server/utils/dbHelpers";
 
 import type { H3Event } from "h3";
 import type { AllowedFileExtensions, DataEntry, AlertsMetadata } from "@/types";
