+++ server/utils/index.ts
@@ -1,25 +1,10 @@
 import type { BasemapConfig, ViewConfig } from "@/types";
-import { validateRowLimit } from "@/utils";
 
 import { fetchTableNames } from "@/server/database/dbOperations";
-
-import { useRuntimeConfig } from "#imports";
-import type { H3Event } from "h3";
 
 export type ParsedBasemaps = {
   basemaps: BasemapConfig[];
   defaultMapboxStyle?: string;
-};
-
-/**
- * Reads the `?limit` query param from an H3 event, validates it against the
- * server row ceiling, and returns the validated limit.
- * Throws an error with `statusCode: 422` for invalid or excessive values.
- */
-export const parseAndValidateLimit = (event: H3Event): number => {
-  const raw = getQuery(event).limit;
-  const maxLimit = Number(useRuntimeConfig(event).public.rowLimit);
-  return validateRowLimit(raw, maxLimit);
 };
 
 /**
